feat(nixl): allocate registerable arenas per device
