+++ src/prime_rl/transports/weights/nixl/device_memory.py
@@ -1,0 +1,112 @@
+"""Device memory that NIXL can register with the NIC."""
+
+from __future__ import annotations
+
+import os
+from contextlib import contextmanager
+from typing import Iterator
+
+import torch
+
+_pool: torch.cuda.MemPool | None = None
+_allocator: torch.cuda.memory.CUDAPluggableAllocator | None = None
+
+
+def _get_pool() -> torch.cuda.MemPool:
+    """Build (once) a cudaMalloc-backed pool, bypassing the caching allocator."""
+    global _pool, _allocator
+    if _pool is not None:
+        return _pool
+
+    import ctypes
+    from pathlib import Path
+
+    from torch.utils.cpp_extension import load_inline
+
+    # The inline extension links against cudart, which is not loaded globally
+    # by default in every environment.
+    try:
+        ctypes.CDLL("libcudart.so", mode=ctypes.RTLD_GLOBAL)
+    except OSError:
+        pass
+
+    module = load_inline(
+        name="cuda_malloc_allocator",
+        cpp_sources=[
+            r"""
+#include <cuda_runtime.h>
+#include <cstddef>
+extern "C" {
+void* cuda_malloc(ptrdiff_t size, int device, void* stream) {
+    (void) stream;
+    int previous = -1;
+    cudaGetDevice(&previous);
+    cudaSetDevice(device);
+    void* pointer = nullptr;
+    cudaError_t error = cudaMalloc(&pointer, (size_t) size);
+    if (previous >= 0) cudaSetDevice(previous);
+    if (error != cudaSuccess) return nullptr;
+    return pointer;
+}
+void cuda_free(void* pointer, ptrdiff_t size, int device, void* stream) {
+    (void) size; (void) stream;
+    int previous = -1;
+    cudaGetDevice(&previous);
+    cudaSetDevice(device);
+    cudaFree(pointer);
+    if (previous >= 0) cudaSetDevice(previous);
+}
+}
+"""
+        ],
+        functions=[],
+        extra_cflags=["-O2"],
+        with_cuda=True,
+    )
+    _allocator = torch.cuda.memory.CUDAPluggableAllocator(str(Path(module.__file__)), "cuda_malloc", "cuda_free")
+    _pool = torch.cuda.MemPool(_allocator.allocator())
+    return _pool
+
+
+def _expandable_segments_enabled(config: str) -> bool:
+    for item in config.split(","):
+        key, separator, value = item.partition(":")
+        if separator and key.strip() == "expandable_segments" and value.strip().lower() == "true":
+            return True
+    return False
+
+
+def _check_caching_allocator_registerable() -> None:
+    """Reject expandable segments, which hand out unregisterable addresses.
+
+    An expandable segment spans several driver-level physical handles, so the NIC
+    cannot pin the virtual range and NIXL registration fails. Both variables are
+    read because a trainer inherits ``PYTORCH_CUDA_ALLOC_CONF`` from
+    ``DEFAULT_TRAINER_ENV_VARS`` regardless of its accelerator.
+    """
+    config = os.environ.get("PYTORCH_ALLOC_CONF")
+    if config is None:
+        config = os.environ.get("PYTORCH_CUDA_ALLOC_CONF", "")
+    if _expandable_segments_enabled(config):
+        raise RuntimeError(
+            "NIXL weight transfer cannot register memory allocated with "
+            "expandable_segments:True. Set PYTORCH_ALLOC_CONF=expandable_segments:False "
+            "in this component's env_vars."
+        )
+
+
+@contextmanager
+def use_registerable_pool(device: torch.device) -> Iterator[None]:
+    """Allocate arenas the NIC can pin, which differs by allocator."""
+    if device.type == "cuda":
+        # The caching allocator runs with expandable segments here, so arenas come
+        # from a dedicated cudaMalloc pool instead.
+        with torch.cuda.use_mem_pool(_get_pool()):
+            yield
+    elif device.type == "xpu":
+        # Registration needs one allocation per region, which the caching allocator
+        # already satisfies once expandable segments are off.
+        _check_caching_allocator_registerable()
+        yield
+    else:
+        raise NotImplementedError(f"NIXL weight transfer does not support {device.type!r} devices")
